+++ flask_sqlalchemy_bundle/__init__.py
@@ -20,6 +20,8 @@ from .decorators import param_converter
 from .extensions import SQLAlchemy, db
 from .model_form import ModelForm
 from .services import ModelManager, SessionManager
+from .validation import (
+    BaseValidator, Required, ValidationError, ValidationErrors, validates)
 
 
 class FlaskSQLAlchemyBundle(Bundle):